iFixes bug with wrong versioning in assemblies
TThere used to be a bug where inserted or derived components would be
updated with the version number of the assembly on save. Now the inserted or
derived component always keep their own version number and this is only
updated when the actual component is updated on save. This is done by
checking if the parentdesign of the component is the same as the
parentdesign of the rootcomponent. If not, this is an inserted/derived
component which should not be updated when saving the assembly.

h with '#' will be ignored, and an empty message aborts the commit.