Update selection to using selectionbox
The script now uses a selection box which means you can select and
unselect as much as you like before clicking ok and stamping. This also
makes it more extendable for future use. Future use could let the user
select what to stamp and how deep or big to stamp.