Adds selection of what to stamp for user
Until now you could only stamp 'Version Number'. This commits adds an
extra selection for the user. You can now choose between Version Number
and alle userparameters. For Version Number the functionality stays the
same. If you choose a user parameter a sketch will be created with it's
name the same as the user parameter. The value of this sketch will then
later be updated by the UpdateTextSketchesOnSave. A possible extension
in the future would be to combine UpdateTextSketchesOnSave and
UpdateVersionTextOnsave into 1 add-in.