add a loop updating a parameter
testing what happens if you add a loop which updates a parameter and
then runs collision detection. This does not seem to work in a script
because the reset to parametric design is postponed by the process
untill after the script. This would probably be fixed by changing it to
a addin with a handler waiting for an event so that the undo can be
executed intermediately. Also the update of the parameter seems to work
but the model does not seem to update, which could have the same cause
als the undo problem.