Testing how to always get the correct path
to the directory of the script. Depending on how the script is run,
sometimes there is an error because the relative path to the resources
folder is not correct. This happened when running from Pycharm. This
update seems to fix it for Windows but it needs to be tested for mac as
well.